+++ CHANGELOG.md
@@ -12,6 +12,7 @@ VERSION v0.11
 - Add `unique` command to be able to merge different bib files
   and filter out repetitions.
 - Add `doctor` to check for fields in a bibfile.
+- Add `iscited` to check which bib items are cited in a text file.
 
 ## Add Format plugin
 - Sébastien Popoff has added a format plugin architecture, so now `jinja2`